+++ eelbrain/test.py
@@ -3,6 +3,7 @@ __test__ = False
 
 from ._stats.glm import anova, ANOVA
 from ._stats.test import (
+    Correlation,
     TTest1Sample, TTestInd, TTestRel, pairwise, ttest, correlations,
     bootstrap_pairwise, lilliefors,
 )